# Conflicts:
#	src/main/java/edu/kosmo/krm/controller/OrderController.java
#	src/main/java/edu/kosmo/krm/mapper/OrderMapper.java
#	src/main/java/edu/kosmo/krm/service/OrderService.java
#	src/main/java/edu/kosmo/krm/service/OrderServiceImpl.java
#	src/main/java/edu/kosmo/krm/service/ProductService.java
#	src/main/resources/mappers/OrderMapper.xml
#	src/main/webapp/WEB-INF/views/user/userHome.jsp

+++ src/main/java/edu/kosmo/krm/joinVO/JoinReviewWriteVO.java
@@ -36,7 +36,7 @@ public class JoinReviewWriteVO {
 	private Timestamp board_date;
 	private int hit;
 	private int score;
-	private int order_id;
+	private Long order_id;
 	private int order_detail_id;
 	private int order_detail_board_id;
 }
